add stats test cases for the paper
can be ran using 'mocha info'. I kept a branch for the version we evaluated and tried the release feature, but it just saves the source code so that's not too useful. i wont be adding anything to the tool besides these tests to confirm the numbers in the paper.