ink
This branch replaces some parts of ncq with ink as enquirer has not been updated for a while. Ink has the benefit fo allowing me to make some cool stuff more easily too (the new save/load system took very little time to program and the underlying display code is not handled by us). This branch just needs to be tested on Windows.